move annotation format functions to data_writer
