+++ ymir/command/mir/tools/data_exporter.py
@@ -5,14 +5,11 @@ exports the assets and annotations from mir format to ark-training-format
 from collections.abc import Collection
 from enum import Enum
 import logging
-import json
 import os
-from typing import Any, Callable, Dict, List, Optional, Set
-import uuid
-import xml.etree.ElementTree as ElementTree
+from typing import Callable, Dict, List, Optional, Set
 
 from mir.protos import mir_command_pb2 as mirpb
-from mir.tools import class_ids, mir_storage_ops
+from mir.tools import class_ids, data_writer, mir_storage_ops
 from mir.tools import utils as mir_utils
 from mir.tools.code import MirCode
 from mir.tools.errors import MirRuntimeError
@@ -61,9 +58,9 @@ def _rel_annotation_path_for_asset(rel_asset_path: str, format_type: ExportForma
 
 def format_file_output_func(anno_format: ExportFormat) -> Callable:
     _format_func_map = {
-        ExportFormat.EXPORT_FORMAT_ARK: _single_image_annotations_to_ark,
-        ExportFormat.EXPORT_FORMAT_VOC: _single_image_annotations_to_voc,
-        ExportFormat.EXPORT_FORMAT_LS_JSON: _single_image_annotations_to_ls_json,
+        ExportFormat.EXPORT_FORMAT_ARK: data_writer._single_image_annotations_to_ark,
+        ExportFormat.EXPORT_FORMAT_VOC: data_writer._single_image_annotations_to_voc,
+        ExportFormat.EXPORT_FORMAT_LS_JSON: data_writer._single_image_annotations_to_ls_json,
     }
     return _format_func_map[anno_format]
 
@@ -289,141 +286,3 @@ def _export_detect_annotations_to_path(asset_ids: List[str], format_type: Export
 
     logging.info(f"missing annotations: {missing_counter}, "
                  f"empty annotations: {empty_counter} out of {len(asset_ids)} assets")
-
-
-def _single_image_annotations_to_ark(asset_id: str, attrs: Any, annotations: List[mirpb.Annotation],
-                                     class_type_mapping: Optional[Dict[int, int]], cls_id_mgr: class_ids.ClassIdManager,
-                                     asset_filename: str) -> str:
-    output_str = ""
-    for annotation in annotations:
-        mapped_id = class_type_mapping[annotation.class_id] if class_type_mapping else annotation.class_id
-        output_str += f"{mapped_id}, {annotation.box.x}, {annotation.box.y}, "
-        output_str += f"{annotation.box.x + annotation.box.w - 1}, {annotation.box.y + annotation.box.h - 1}\n"
-    return output_str
-
-
-def _single_image_annotations_to_voc(asset_id: str, attrs: Any, annotations: List[mirpb.Annotation],
-                                     class_type_mapping: Optional[Dict[int, int]], cls_id_mgr: class_ids.ClassIdManager,
-                                     asset_filename: str) -> str:
-    # annotation
-    annotation_node = ElementTree.Element('annotation')
-
-    # annotation: folder
-    folder_node = ElementTree.SubElement(annotation_node, 'folder')
-    folder_node.text = 'folder'
-
-    # annotation: filename
-    filename_node = ElementTree.SubElement(annotation_node, 'filename')
-    filename_node.text = asset_filename
-
-    # annotation: source
-    source_node = ElementTree.SubElement(annotation_node, 'source')
-
-    # annotation: source: database
-    database_node = ElementTree.SubElement(source_node, 'database')
-    database_node.text = attrs.dataset_name or 'unknown'
-
-    # annotation: source: annotation
-    annotation2_node = ElementTree.SubElement(source_node, 'annotation')
-    annotation2_node.text = 'unknown'
-
-    # annotation: source: image
-    image_node = ElementTree.SubElement(source_node, 'image')
-    image_node.text = 'unknown'
-
-    # annotation: size
-    size_node = ElementTree.SubElement(annotation_node, 'size')
-
-    # annotation: size: width
-    width_node = ElementTree.SubElement(size_node, 'width')
-    width_node.text = str(attrs.width)
-
-    # annotation: size: height
-    height_node = ElementTree.SubElement(size_node, 'height')
-    height_node.text = str(attrs.height)
-
-    # annotation: size: depth
-    depth_node = ElementTree.SubElement(size_node, 'depth')
-    depth_node.text = str(attrs.image_channels)
-
-    # annotation: segmented
-    segmented_node = ElementTree.SubElement(annotation_node, 'segmented')
-    segmented_node.text = '0'
-
-    # annotation: object(s)
-    for annotation in annotations:
-        object_node = ElementTree.SubElement(annotation_node, 'object')
-
-        name_node = ElementTree.SubElement(object_node, 'name')
-        name_node.text = cls_id_mgr.main_name_for_id(annotation.class_id) or 'unknown'
-
-        pose_node = ElementTree.SubElement(object_node, 'pose')
-        pose_node.text = 'unknown'
-
-        truncated_node = ElementTree.SubElement(object_node, 'truncated')
-        truncated_node.text = 'unknown'
-
-        occluded_node = ElementTree.SubElement(object_node, 'occluded')
-        occluded_node.text = '0'
-
-        bndbox_node = ElementTree.SubElement(object_node, 'bndbox')
-
-        xmin_node = ElementTree.SubElement(bndbox_node, 'xmin')
-        xmin_node.text = str(annotation.box.x)
-
-        ymin_node = ElementTree.SubElement(bndbox_node, 'ymin')
-        ymin_node.text = str(annotation.box.y)
-
-        xmax_node = ElementTree.SubElement(bndbox_node, 'xmax')
-        xmax_node.text = str(annotation.box.x + annotation.box.w - 1)
-
-        ymax_node = ElementTree.SubElement(bndbox_node, 'ymax')
-        ymax_node.text = str(annotation.box.y + annotation.box.h - 1)
-
-        difficult_node = ElementTree.SubElement(object_node, 'difficult')
-        difficult_node.text = '0'
-
-    return ElementTree.tostring(element=annotation_node, encoding='unicode')
-
-
-def _single_image_annotations_to_ls_json(asset_id: str, attrs: Any, annotations: List[mirpb.Annotation],
-                                         class_type_mapping: Optional[Dict[int, int]],
-                                         cls_id_mgr: class_ids.ClassIdManager, asset_filename: str) -> str:
-    out_type = "predictions"  # out_type: annotation type - "annotations" or "predictions"
-    to_name = 'image'  # to_name: object name from Label Studio labeling config
-    from_name = 'label'  # control tag name from Label Studio labeling config
-    task: Dict = {
-        out_type: [{
-            "result": [],
-            "ground_truth": False,
-        }],
-        "data": {
-            "image": asset_filename
-        }
-    }
-
-    for annotation in annotations:
-        bbox_x, bbox_y = float(annotation.box.x), float(annotation.box.y)
-        bbox_width, bbox_height = float(annotation.box.w), float(annotation.box.h)
-        img_width, img_height = attrs.width, attrs.height
-        item = {
-            "id": uuid.uuid4().hex[0:10],  # random id to identify this annotation.
-            "type": "rectanglelabels",
-            "value": {
-                # Units of image annotations in label studio is percentage of image width/height.
-                # https://labelstud.io/guide/predictions.html#Units-of-image-annotations
-                "x": bbox_x / img_width * 100,
-                "y": bbox_y / img_height * 100,
-                "width": bbox_width / img_width * 100,
-                "height": bbox_height / img_height * 100,
-                "rotation": 0,
-                "rectanglelabels": [cls_id_mgr.main_name_for_id(annotation.class_id) or 'unknown']
-            },
-            "to_name": to_name,
-            "from_name": from_name,
-            "image_rotation": 0,
-            "original_width": img_width,
-            "original_height": img_height
-        }
-        task[out_type][0]['result'].append(item)
-    return json.dumps(task)
